pyproject.toml
[tool.poetry]
name = "freva"
version = "0.1.0"
description = ""
authors = ["Andrej Fast <andrej.fast@dkrz.de>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
mysqlclient = "^2.1.1"
django-webpack-loader = "^1.7.0"
requests = "^2.28.1"
pydantic = "^1.10.2"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
mypy = "^0.991"
pytest = "^7.2.0"
pylint = "^2.15.5"
# For reasons which are unclear, this is a fork of django-stubs which offers slightly different types and vscode ships
# with and uses these instead of django-stubs even if it's installed. Rather than deal with different tools offering
# different answers about what types things are, I'm having it just use these so hopefully make this all uniform
django-types = "^0.16.0"
# this has a similar relation to djagnorestframework-stubs
djangorestframework-types = "^0.8.0"
pyright = "^1.1.281"

[tool.pyright]
typeCheckingMode = "strict"
# reportPrivateUsage = "warning"
# reportUnusedImport = "warning"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
